Make test helper classes more consistent.
