+++ src/DotLiquid.Tests/HashTests.cs
@@ -25,24 +25,28 @@ namespace DotLiquid.Tests
     [TestFixture]
     public class HashTests
     {
-        public class TestBaseClass
+        #region Classes used in tests
+
+        private class TestBaseClass
         {
             public string TestBaseClassProp { get; set; }
 
             public virtual string TestOverridableProp { get; set; }
         }
 
-        public class TestMiddleClass : TestBaseClass
+        private class TestMiddleClass : TestBaseClass
         {
             public string TestMiddleClassProp { get; set; }
         }
 
-        public class TestChildClass : TestMiddleClass
+        private class TestChildClass : TestMiddleClass
         {
             public string TestClassProp { get; set; }
 
             public override string TestOverridableProp { get; set; }
         }
+
+        #endregion
 
         #region Mapper Cache Tests
         /// <summary>
